==> 6-google-storage.py <==
from utils.send import sendToGoogleStorage

sendToGoogleStorage('2-cwns-data-distance', 'cwns-data-distance')
sendToGoogleStorage('3-power-data-distance', 'power-data-distance') 
sendToGoogleStorage('4-transport', 'transport')
sendToGoogleStorage('5-cwns-data-distance-incl-transport', 'cwns-data-distance-incl-transport')
sendToGoogleStorage('5-power-data-distance-incl-transport', 'power-data-distance-incl-transport')
